requirements.txt
boto3==1.40.68
botocore==1.40.68
certifi==2025.10.5
charset-normalizer==3.4.4
filelock==3.19.1
fsspec==2025.9.0
h5py==3.14.0
hf-xet==1.2.0
huggingface-hub==0.36.0
idna==3.11
Jinja2==3.1.6
jmespath==1.0.1
joblib==1.5.2
MarkupSafe==2.1.5
mpmath==1.3.0
networkx==3.2.1
numpy==1.26.3
packaging==25.0
pillow==11.3.0
psycopg2-binary==2.9.11
python-dateutil==2.9.0.post0
PyYAML==6.0.3
regex==2025.11.3
requests==2.32.5
s3transfer==0.14.0
safetensors==0.6.2
scikit-learn==1.6.1
scipy==1.13.1
sentence-transformers==2.7.0
six==1.17.0
sympy==1.14.0
threadpoolctl==3.6.0
tokenizers==0.22.1
torch==2.4.1+cpu
torchaudio==2.4.1+cpu
torchvision==0.19.1+cpu
tqdm==4.67.1
transformers==4.57.1
typing_extensions==4.15.0
urllib3==1.26.20
